+++ tests/index.php
@@ -1,4 +1,12 @@
 <?php
+/**
+ * Tests: References
+ * 
+ * - [PHPUnit](https://github.com/sebastianbergmann/phpunit)
+ * - [PHPUnit Polyfills](https://github.com/Yoast/PHPUnit-Polyfills)
+ * - [WP_UnitTestCase](https://github.com/WordPress/wordpress-develop/blob/trunk/tests/phpunit/includes/abstract-testcase.php)
+ * - [Assertions](https://docs.phpunit.de/en/10.2/assertions.html)
+ */
 
 if ( ! $_WORDPRESS_DEVELOP_DIR = getenv( 'WORDPRESS_DEVELOP_DIR' ) ) {
     $_WORDPRESS_DEVELOP_DIR = __DIR__ . '/../wordpress-develop';