+++ docs/changelog.md
@@ -1,5 +1,11 @@
 Versions correspond to plugin release of Loops & Logic and Tangible Blocks.
 
+# 3.2.9
+
+- Format tag
+  - Format list index: Fix warning from array_shift()
+  - Support capture groups for replace_pattern, for example: replace_pattern="/(\d{3})/" with="$1"
+  - When invalid regular expression is passed, emit a warning instead of throwing an error
 
 # 3.2.8